allow float n in rust

+++ src/derivatives.rs
@@ -13,8 +13,8 @@ pub fn get_yprime() -> &'static dyn Fn(f64,f64,f64)->f64 {
 //     }
 //     &zprime
 // }
-pub fn get_zprime(n: usize) -> Box<dyn Fn(f64, f64, f64) -> f64> {
+pub fn get_zprime(n: f64) -> Box<dyn Fn(f64, f64, f64) -> f64> {
     Box::new(move |x, y, z| {
-        -1.0 * y.powf(n as f64) - 2.0 / x * z
+        -1.0 * y.powf(n) - 2.0 / x * z
     })
 }
